+++ lib/tests/test_battery_model.py
@@ -1,17 +1,19 @@
 from lib.battery_model import Battery
 
+import numpy as np
+
 
 def test_battery_cycle():
     battery = Battery(
-        soc_0=0.5,
-        minimum_soc=0.0,
-        efficiency=1.0,
-        maximum_energy=1.0,
-        maximum_power=1e9,
+        soc_0=np.float64(0.5),
+        minimum_soc=np.float64(0.0),
+        efficiency=np.float64(1.0),
+        maximum_energy=np.float64(1.0),
+        maximum_power=np.float64(1e9),
     )
 
-    dt = 0.5
-    power = 1.0
+    dt = np.float64(0.5)
+    power = np.float64(1.0)
     soc_before_cycle = battery.soc
 
     for i in range(10):
@@ -26,28 +28,28 @@ def test_battery_cycle():
 
 
 def test_battery_energy_limits():
-    maximum_energy = 1.0
-    soc_0 = 1.0
-    minimum_soc = 0.5
+    maximum_energy = np.float64(1.0)
+    soc_0 = np.float64(1.0)
+    minimum_soc = np.float64(0.5)
     battery = Battery(
         soc_0=soc_0,
         minimum_soc=minimum_soc,
-        efficiency=1.0,
+        efficiency=np.float64(1.0),
         maximum_energy=maximum_energy,
-        maximum_power=1e9,
+        maximum_power=np.float64(1e9),
     )
 
-    dt = 1.0
+    dt = np.float64(1.0)
 
     for i in range(10):
         # Discharges half of the battery in one second
-        power = (soc_0 - minimum_soc) * maximum_energy / (dt / 3600)
+        power = (soc_0 - minimum_soc) * maximum_energy / (dt / np.float64(3600))
         assert battery.solve(dt, -power) == -power
         assert battery.soc == minimum_soc
 
         for i in range(10):
             # Should fail to discharge the remaining of the battery
-            assert battery.solve(dt, -power) == 0.0
+            assert battery.solve(dt, -power) == np.float64(0.0)
             assert battery.soc == minimum_soc
 
         # Now if we recharge it, should go back to the initial state
@@ -56,22 +58,22 @@ def test_battery_energy_limits():
 
         for i in range(10):
             # And finally, it should not charge beyond full
-            assert battery.solve(dt, power) == 0.0
+            assert battery.solve(dt, power) == np.float64(0.0)
             assert battery.soc == soc_0
 
 
 def test_battery_power_limit():
-    maximum_power = 5.0
+    maximum_power = np.float64(5.0)
     battery = Battery(
-        soc_0=0.5,
-        minimum_soc=0.0,
-        efficiency=1.0,
-        maximum_energy=1.0,
+        soc_0=np.float64(0.5),
+        minimum_soc=np.float64(0.0),
+        efficiency=np.float64(1.0),
+        maximum_energy=np.float64(1.0),
         maximum_power=maximum_power,
     )
 
-    dt = 1.0
-    power = 10.0
+    dt = np.float64(1.0)
+    power = np.float64(10.0)
 
     for i in range(10):
         # Out of the bounds of maximum power should return a limited power
@@ -81,31 +83,32 @@ def test_battery_power_limit():
 
 def test_battery_charge():
     battery = Battery(
-        soc_0=0.0,
-        minimum_soc=0.0,
-        efficiency=1.0,
-        maximum_energy=1.0,
-        maximum_power=1e9,
+        soc_0=np.float64(0.0),
+        minimum_soc=np.float64(0.0),
+        efficiency=np.float64(1.0),
+        maximum_energy=np.float64(1.0),
+        maximum_power=np.float64(1e9),
     )
 
-    duration = 0.5  # in hours
-    dt = duration * 3600 # in seconds
+    duration = np.float64(0.5)  # in hours
+    dt = duration * np.float64(3600)  # in seconds
     power = battery.maximum_energy / duration
 
     assert battery.solve(dt, power) == power
-    assert battery.soc == 1.0
+    assert battery.soc == np.float64(1.0)
 
 
 def test_battery_discharge():
     battery = Battery(
-        soc_0=1.0,
-        minimum_soc=0.0,
-        efficiency=1.0,
-        maximum_energy=1.0,
-        maximum_power=1e9,    )
+        soc_0=np.float64(1.0),
+        minimum_soc=np.float64(0.0),
+        efficiency=np.float64(1.0),
+        maximum_energy=np.float64(1.0),
+        maximum_power=np.float64(1e9),
+    )
 
-    duration = 0.5  # in hours
-    dt = duration * 3600) # in seconds
+    duration = np.float64(0.5)  # in hours
+    dt = duration * np.float64(3600)  # in seconds
     power = -battery.maximum_energy / duration
 
     assert battery.solve(dt, power) == power
